backend/requirements.txt
fastapi
pydantic
uvicorn
tortoise-orm[asyncpg]
python-multipart
